Add types for Link
